requirements.txt
anthropic>=0.40.0
cryptography>=42.0.0
slack-bolt>=1.18.0
slack-sdk>=3.27.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
sentry-sdk>=2.0.0
pydantic>=2.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
